x permissions removed. pip do not preserve the x flag. Some test rescued, moved to pyworkflowtests.
