+++ pyworkflow/gui/canvas.py
