+++ pyworkflow/gui/graph.py
