+++ pyworkflow/gui/project/viewdata.py
