+++ pyworkflow/gui/project/viewprotocols.py
