+++ pyworkflow/protocol/protocol.py
@@ -2177,11 +2177,11 @@ def runProtocolMain(projectPath, protDbPath, protId):
             # We run "scipion run pyworkflow/...mpirun.py blah" instead of
             # calling directly "$SCIPION_PYTHON ...mpirun.py blah", so that
             # when it runs on a MPI node, it *always* has the scipion env.
-            params = ['runprotocol', pw.getPwProtMpiRunScript(),
+            params = ['-m', 'scipion', 'runprotocol', pw.getPwProtMpiRunScript(),
                       projectPath, protDbPath, protId]
             # 'scipion' is treated now as an entry point, but if there is an alias with that name, the alias has higher
             # priority
-            retcode = pwutils.runJob(None, 'scipion', params,
+            retcode = pwutils.runJob(None, pw.PYTHON, params,
                                      numberOfMpi=protocol.numberOfMpi.get(),
                                      hostConfig=hostConfig)
             sys.exit(retcode)
