+++ pyworkflow/resources/class_obj.gif
