+++ pyworkflow/resources/file_sqlite.gif
