+++ pyworkflow/resources/file_stack.gif
